+++ src/com/anji/integration/ConsoleLogEventListener.java
@@ -73,10 +73,10 @@ public void geneticEventFired(GeneticEvent event) {
 	if ( GeneticEvent.GENOTYPE_EVOLVED_EVENT.equals( event.getEventName() ) ) {
 		Genotype genotype = (Genotype) event.getSource();
 		Chromosome fittest = genotype.getFittestChromosome();
-		float maxFitnessValue = (config.getBulkFitnessFunction() != null) ?
+		double maxFitnessValue = (config.getBulkFitnessFunction() != null) ?
 			config.getBulkFitnessFunction().getMaxFitnessValue() :
 			config.getFitnessFunction().getMaxFitnessValue();
-		float fitness = (maxFitnessValue == 0) ? fittest.getFitnessValue() :
+		double fitness = (maxFitnessValue == 0) ? fittest.getFitnessValue() :
 			(fittest.getFitnessValue() / maxFitnessValue);
 		out.println( "species count: " + genotype.getSpecies().size() );
 		out.println( "fittest chromosome: " + fittest.getId() + ", score == " + fitness + " and # genes == " + fittest.size() );
