+++ src/com/anji/neat/AddConnectionMutationOperator.java
@@ -56,7 +56,7 @@ public class AddConnectionMutationOperator extends MutationOperator implements C
 /**
  * default mutation rate
  */
-public static final float DEFAULT_MUTATE_RATE = 0.01f;
+public static final double DEFAULT_MUTATE_RATE = 0.01f;
 
 private RecurrencyPolicy policy;
 
@@ -64,13 +64,13 @@ private RecurrencyPolicy policy;
  * @see com.anji.util.Configurable#init(com.anji.util.Properties)
  */
 public void init( Properties props ) throws Exception {
-	setMutationRate( props.getFloatProperty( ADD_CONN_MUTATE_RATE_KEY,
+	setMutationRate( props.getDoubleProperty( ADD_CONN_MUTATE_RATE_KEY,
 			DEFAULT_MUTATE_RATE ) );
 	policy = RecurrencyPolicy.load( props );
 }
 
 /**
- * @see AddConnectionMutationOperator#AddConnectionMutationOperator(float)
+ * @see AddConnectionMutationOperator#AddConnectionMutationOperator(double)
  */
 public AddConnectionMutationOperator() {
 	this( DEFAULT_MUTATE_RATE, RecurrencyPolicy.BEST_GUESS );
@@ -78,9 +78,9 @@ public AddConnectionMutationOperator() {
 
 /**
  * @param newMutationRate
- * @see AddConnectionMutationOperator#AddConnectionMutationOperator(float, RecurrencyPolicy)
+ * @see AddConnectionMutationOperator#AddConnectionMutationOperator(double, RecurrencyPolicy)
  */
-public AddConnectionMutationOperator( float newMutationRate ) {
+public AddConnectionMutationOperator( double newMutationRate ) {
 	this( newMutationRate, RecurrencyPolicy.BEST_GUESS );
 }
 
@@ -101,7 +101,7 @@ public AddConnectionMutationOperator( RecurrencyPolicy aPolicy ) {
  * @param aPolicy
  * @see RecurrencyPolicy
  */
-public AddConnectionMutationOperator( float aMutationRate, RecurrencyPolicy aPolicy ) {
+public AddConnectionMutationOperator( double aMutationRate, RecurrencyPolicy aPolicy ) {
 	super( aMutationRate );
 	policy = aPolicy;
 }
